Updated given changes regarding custom class, naming and commenting conventions.
